response nothing is playing right now
